worker/requirements.txt
python-dotenv>=1.0.0
aiohttp>=3.10.0
click>=8.1.0
requests>=2.31.0
pytest>=7.4.0
pytest-asyncio>=0.21.1
slack_sdk>=3.24.0

typing-extensions>=4.5.0 

portkey-ai>=0.10.0
elevenlabs>=0.1.0

# Utilities
python-json-logger>=2.0.7
pydantic>=2.0.0
requests>=2.31.0
websockets>=11.0.3